Move ModuleAlias widget classes to src/hub/module_alias.hh
